Quickstart update
#1 Joomla update 3.8.3 to 3.8.8
#2 Kunena Forum update 5.0 to 5.1.1
#3 Virtuemart update in latest version 3.2.14
#4 Update ChronoForms latest version V6.0.18
#5 Remove "position : relative" From off-canvas-menu-init class in menu.less.
#6 Change version in template xml file 2.1 to 2.2

+++ quickstart/administrator/components/com_virtuemart/views/orders/view.html.php
@@ -5,7 +5,7 @@
  *
  * @package	VirtueMart
  * @subpackage
- * @author
+ * @author Max Milbers, Valérie Isaksen
  * @link https://virtuemart.net
  * @copyright Copyright (c) 2004 - 2010 VirtueMart Team. All rights reserved.
  * @license http://www.gnu.org/copyleft/gpl.html GNU/GPL, see LICENSE.php
@@ -73,7 +73,8 @@ class VirtuemartViewOrders extends VmViewAdmin {
 			$_orderID = $order['details']['BT']->virtuemart_order_id;
 			$orderbt = $order['details']['BT'];
 			$orderst = $order['details']['ST'];//(array_key_exists('ST', $order['details'])) ? $order['details']['ST'] : $orderbt;
-			$orderbt ->invoiceNumber = $orderModel->getInvoiceNumber($orderbt->virtuemart_order_id);
+			$invM = VmModel::getModel('invoice');
+			$orderbt ->invoiceNumbers = $invM->getInvoiceNumbers($orderbt->virtuemart_order_id);
 
 			$currency = CurrencyDisplay::getInstance(0,$order['details']['BT']->virtuemart_vendor_id);
 
@@ -125,8 +126,11 @@ class VirtuemartViewOrders extends VmViewAdmin {
 
 			$_itemStatusUpdateFields = array();
 			$_itemAttributesUpdateFields = array();
+			// for the order item template
+			$_itemStatusUpdateFields[0] = JHtml::_('select.genericlist', $orderStates, "item_id[0][order_status]", 'class="selectItemStatusCode"', 'order_status_code', 'order_status_name', 'P', 'order_item_status_0',true);
+
 			foreach($order['items'] as $_item) {
-				$_itemStatusUpdateFields[$_item->virtuemart_order_item_id] = JHtml::_('select.genericlist', $orderStates, "item_id[".$_item->virtuemart_order_item_id."][order_status]", 'class="selectItemStatusCode"', 'order_status_code', 'order_status_name', $_item->order_status, 'order_item_status'.$_item->virtuemart_order_item_id,true);
+				$_itemStatusUpdateFields[$_item->virtuemart_order_item_id] = JHtml::_('select.genericlist', $orderStates, "item_id[".$_item->virtuemart_order_item_id."][order_status]", 'class="selectItemStatusCode" style="width:160px;', 'order_status_code', 'order_status_name', $_item->order_status, 'order_item_status'.$_item->virtuemart_order_item_id,true);
 
 				$_item->linkedit = 'index.php?option=com_virtuemart&view=product&task=edit&virtuemart_product_id='.$_item->virtuemart_product_id;
 			}
@@ -155,7 +159,7 @@ class VirtuemartViewOrders extends VmViewAdmin {
 			/* Data for the Edit Status form popup */
 			$_currentOrderStat = $order['details']['BT']->order_status;
 			// used to update all item status in one time
-			$_orderStatusSelect = JHtml::_('select.genericlist', $orderStates, 'order_status', 'style="width:100px;"', 'order_status_code', 'order_status_name', $_currentOrderStat, 'order_items_status',true);
+			$_orderStatusSelect = JHtml::_('select.genericlist', $orderStates, 'order_status', 'style="width:200px;', 'order_status_code', 'order_status_name', $_currentOrderStat, 'order_items_status',true);
 			$this->assignRef('orderStatSelect', $_orderStatusSelect);
 			$this->assignRef('currentOrderStat', $_currentOrderStat);
 
@@ -210,7 +214,7 @@ class VirtuemartViewOrders extends VmViewAdmin {
 			$_currencies = array(); // Save the currency data during this loop for performance reasons
 
 			if ($orderslist) {
-
+				$invM = VmModel::getModel('invoice');
 			    foreach ($orderslist as $virtuemart_order_id => $order) {
 
 				    if(!empty($order->order_currency)){
@@ -227,7 +231,8 @@ class VirtuemartViewOrders extends VmViewAdmin {
 				    }
 
 					$orderslist[$virtuemart_order_id]->order_total = $_currencies['curr'.$currency]->priceDisplay($order->order_total);
-					$orderslist[$virtuemart_order_id]->invoiceNumber = $model->getInvoiceNumber($order->virtuemart_order_id);
+
+					$orderslist[$virtuemart_order_id]->invoiceNumbers = $invM->getInvoiceNumbers($order->virtuemart_order_id);
 			    }
 
 			}
@@ -252,7 +257,7 @@ class VirtuemartViewOrders extends VmViewAdmin {
 
 			JToolBarHelper::save('updatestatus', vmText::_('COM_VIRTUEMART_UPDATE_STATUS'));
 
-			if (vmAccess::manager('orders.delete')) {
+			if (vmAccess::manager('orders.delete') && !VmConfig::get('ordersAddOnly',false)) {
 				JToolBarHelper::spacer('80');
 				JToolBarHelper::deleteList();
 			}
@@ -273,32 +278,52 @@ class VirtuemartViewOrders extends VmViewAdmin {
 		parent::display($tpl);
 	}
 
-	function createPrintLinks($order,&$print_link,&$deliverynote_link,&$invoice_link){
+
+	/**
+	 * @author Max Milbers
+	 * @author Valérie Isaksen
+	 * @param $order
+	 * @param $print_link
+	 * @param $deliverynote_link
+	 * @param $invoice_links
+	 */
+	function createPrintLinks($order,&$print_link,&$deliverynote_link,&$invoice_links){
 
 		$baseUrl = 'index.php?option=com_virtuemart&view=orders&task=callInvoiceView&tmpl=component&virtuemart_order_id=' . $order->virtuemart_order_id;
 		/* Print view URL */
 		$print_url = $baseUrl .'&layout=invoice';
 		$print_link = "<a href=\"javascript:void window.open('$print_url', 'win2', 'status=no,toolbar=no,scrollbars=yes,titlebar=no,menubar=no,resizable=yes,width=640,height=480,directories=no,location=no');\"  >";
 		$print_link .= '<span class="hasTip print_32" title="' . vmText::_ ('COM_VIRTUEMART_PRINT').' '. $order->order_number.'">&nbsp;</span></a>';
-		$invoice_link = '';
+		$invoice_links_array = array();
 		$deliverynote_link = '';
 		$pdfDummi= '&d='.rand(0,100);
-		if (!$order->invoiceNumber) {
+		if (!$order->invoiceNumbers) {
 			$invoice_url = $baseUrl .'&layout=invoice&format=pdf&create_invoice='.$order->order_create_invoice_pass.$pdfDummi;
-			$invoice_link .= "<a href=\"$invoice_url\"  >".'<span class="hasTip invoicenew_32" title="' . vmText::_ ('COM_VIRTUEMART_INVOICE_CREATE') . '"></span></a>';
-		} elseif (!shopFunctions::InvoiceNumberReserved ($order->invoiceNumber)) {
-			$invoice_url = $baseUrl .'&layout=invoice&format=pdf'.$pdfDummi;
-			$invoice_link = "<a href=\"$invoice_url\"  >" . '<span class="hasTip invoice_32" title="' . vmText::_ ('COM_VIRTUEMART_INVOICE') .' '.$order->invoiceNumber. '"></span></a>';
+			$invoice_links_array[]= "<a href=\"$invoice_url\"  >".'<span class="hasTip invoicenew_32" title="' . vmText::_ ('COM_VIRTUEMART_INVOICE_CREATE') . '"></span></a>';
+		} else {
+			foreach ($order->invoiceNumbers as $invoiceNumber) {
+				if (!shopFunctions::InvoiceNumberReserved ($invoiceNumber)) {
+					$invoice_url = $baseUrl .'&layout=invoice&format=pdf'.$pdfDummi.'&invoiceNumber='.$invoiceNumber;
+					$invoice_links_array[] = "<a href=\"$invoice_url\"  >" . '<span class="hasTip invoice_32" title="' . vmText::_ ('COM_VIRTUEMART_INVOICE') .' '.$invoiceNumber. '"></span></a>';
+				}
+			}
 		}
 
-		if (!$order->invoiceNumber) {
+		if (!$order->invoiceNumbers) {
 			$deliverynote_url = $baseUrl .'&layout=deliverynote&format=pdf&create_invoice='.$order->order_create_invoice_pass.$pdfDummi;
 			$deliverynote_link = "<a href=\"$deliverynote_url\"  >" . '<span class="hasTip deliverynotenew_32" title="' . vmText::_ ('COM_VIRTUEMART_DELIVERYNOTE_CREATE') . '"></span></a>';
-		} elseif (!shopFunctionsF::InvoiceNumberReserved ($order->invoiceNumber)) {
-			$deliverynote_url = $baseUrl .'&layout=deliverynote&format=pdf&virtuemart_order_id=' . $order->virtuemart_order_id .$pdfDummi;
-			$deliverynote_link = "<a href=\"$deliverynote_url\"  >" . '<span class="hasTip deliverynote_32" title="' . vmText::_ ('COM_VIRTUEMART_DELIVERYNOTE').' '.$order->invoiceNumber . '"></span></a>';
+		} else {
+			/*
+			 * TODO: InvoiceNumberReserved is used by some payments like Klarna
+			 */
+			$invoiceNumber= $order->invoiceNumbers [0];
+			if (!shopFunctionsF::InvoiceNumberReserved ($invoiceNumber)) {
+				$deliverynote_url = $baseUrl .'&layout=deliverynote&format=pdf&virtuemart_order_id=' . $order->virtuemart_order_id .$pdfDummi;
+				$deliverynote_link = "<a href=\"$deliverynote_url\"  >" . '<span class="hasTip deliverynote_32" title="' . vmText::_ ('COM_VIRTUEMART_DELIVERYNOTE').' '.$invoiceNumber . '"></span></a>';
 		}
 
+		}
+		$invoice_links=implode(' ', $invoice_links_array);
 	}
 }
